[issue-159] Move Browser update logic into component
